# server/controllers/SchoolMS_db/generated/TestControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE TestControllerGenerated.js PLEASE EDIT ../TestController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import TestModel from "../../../models/SchoolMS_db/TestModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import TestController from "../TestController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/test`;
    router.post(baseUrl + "", authorize([]), TestController.create);
    router.delete(baseUrl + "/:id", authorize([]), TestController.delete);
    router.get(baseUrl + "/findBy_Class/:key", authorize([]), TestController.findBy_Class);
    router.get(baseUrl + "/findBy_Lesson/:key", authorize([]), TestController.findBy_Lesson);
    router.get(baseUrl + "/findBy_Subject/:key", authorize([]), TestController.findBy_Subject);
    router.get(baseUrl + "/findBy_Teacher/:key", authorize([]), TestController.findBy_Teacher);
    router.get(baseUrl + "/findByclass/:key", authorize([]), TestController.findByclass);
    router.get(baseUrl + "/findBydate/:key", authorize([]), TestController.findBydate);
    router.get(baseUrl + "/findBylessonno/:key", authorize([]), TestController.findBylessonno);
    router.get(baseUrl + "/findBysubjectname/:key", authorize([]), TestController.findBysubjectname);
    router.get(baseUrl + "/findByteachername/:key", authorize([]), TestController.findByteachername);
    router.get(baseUrl + "/:id", authorize([]), TestController.get);
    router.get(baseUrl + "/:id/get_Class", authorize([]), TestController.get_Class);
    router.get(baseUrl + "/:id/get_Lesson", authorize([]), TestController.get_Lesson);
    router.get(baseUrl + "/:id/get_Subject", authorize([]), TestController.get_Subject);
    router.get(baseUrl + "/:id/get_Teacher", authorize([]), TestController.get_Teacher);
    router.get(baseUrl + "", authorize([]), TestController.list);
    router.post(baseUrl + "/:id", authorize([]), TestController.update);
  },


  // CRUD METHODS


  /**
  * TestModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await TestModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.delete
  *   @description CRUD ACTION delete
  *   @param Number class
  *   @param ObjectId id Id
  *   @param Number lessonno
  *   @param String subjectname
  *
  */
  delete: async (req, res) => {
    try {
      const result = await TestModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.findBy_Class
  *   @description CRUD ACTION findBy_Class
  *
  */
  findBy_Class: async (req, res) => {
    try {
      const result = await TestModel.findBy_Class(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.findBy_Lesson
  *   @description CRUD ACTION findBy_Lesson
  *
  */
  findBy_Lesson: async (req, res) => {
    try {
      const result = await TestModel.findBy_Lesson(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.findBy_Subject
  *   @description CRUD ACTION findBy_Subject
  *
  */
  findBy_Subject: async (req, res) => {
    try {
      const result = await TestModel.findBy_Subject(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.findBy_Teacher
  *   @description CRUD ACTION findBy_Teacher
  *
  */
  findBy_Teacher: async (req, res) => {
    try {
      const result = await TestModel.findBy_Teacher(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.findByclass
  *   @description CRUD ACTION findByclass
  *
  */
  findByclass: async (req, res) => {
    try {
      const result = await TestModel.findByclass(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.findBydate
  *   @description CRUD ACTION findBydate
  *
  */
  findBydate: async (req, res) => {
    try {
      const result = await TestModel.findBydate(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.findBylessonno
  *   @description CRUD ACTION findBylessonno
  *
  */
  findBylessonno: async (req, res) => {
    try {
      const result = await TestModel.findBylessonno(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.findBysubjectname
  *   @description CRUD ACTION findBysubjectname
  *
  */
  findBysubjectname: async (req, res) => {
    try {
      const result = await TestModel.findBysubjectname(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.findByteachername
  *   @description CRUD ACTION findByteachername
  *
  */
  findByteachername: async (req, res) => {
    try {
      const result = await TestModel.findByteachername(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.get
  *   @description CRUD ACTION get
  *   @param Number class
  *   @param ObjectId id Id resource
  *   @param Number lessonno
  *   @param String subjectname
  *
  */
  get: async (req, res) => {
    try {
      const result = await TestModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.get_Class
  *   @description CRUD ACTION get_Class
  *   @param Number class
  *
  */
  get_Class: async (req, res) => {
    try {
      const result = await TestModel.get_Class(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.get_Lesson
  *   @description CRUD ACTION get_Lesson
  *   @param String lessonname
  *
  */
  get_Lesson: async (req, res) => {
    try {
      const result = await TestModel.get_Lesson(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.get_Subject
  *   @description CRUD ACTION get_Subject
  *   @param String subjectname
  *
  */
  get_Subject: async (req, res) => {
    try {
      const result = await TestModel.get_Subject(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.get_Teacher
  *   @description CRUD ACTION get_Teacher
  *   @param String teachername
  *
  */
  get_Teacher: async (req, res) => {
    try {
      const result = await TestModel.get_Teacher(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * TestModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await TestModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * TestModel.update
  *   @description CRUD ACTION update
  *   @param Number class
  *   @param ObjectId id Id
  *   @param Number lessonno
  *   @param String subjectname
  *
  */
  update: async (req, res) => {
    try {
      const result = await TestModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;