# server/swagger.yaml
openapi: 3.0.0
info:
  title: SchoolMS
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /test/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Test
      parameters:
        - name: class
          in: path
          description: 
          required: true
          schema:
            type: number
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: lessonno
          in: path
          description: 
          required: true
          schema:
            type: number
        - name: subjectname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Test
      parameters:
        - name: class
          in: path
          description: 
          required: true
          schema:
            type: number
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
        - name: lessonno
          in: path
          description: 
          required: true
          schema:
            type: number
        - name: subjectname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
    post:
      summary: CRUD ACTION update
      tags:
        - Test
      parameters:
        - name: class
          in: path
          description: 
          required: true
          schema:
            type: number
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: lessonno
          in: path
          description: 
          required: true
          schema:
            type: number
        - name: subjectname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Test"
...
  /test/{id}/get_Class:
    get:
      summary: CRUD ACTION get_Class
      tags:
        - Test
      parameters:
        - name: class
          in: path
          description: 
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
  /test/{id}/get_Lesson:
    get:
      summary: CRUD ACTION get_Lesson
      tags:
        - Test
      parameters:
        - name: lessonname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /test/{id}/get_Subject:
    get:
      summary: CRUD ACTION get_Subject
      tags:
        - Test
      parameters:
        - name: subjectname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /test/{id}/get_Teacher:
    get:
      summary: CRUD ACTION get_Teacher
      tags:
        - Test
      parameters:
        - name: teachername
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  User  -----
...
  /user/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
...
  /Users/{id}/changePassword:
    post:
      summary: Change password by admin API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordAdmin:
                  type: string
      responses:
        '200':
          description: OK
  /Users/{id}:
    post:
      summary: Update User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    delete:
      summary: Delete User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Test:
      properties:
        _id:
          type: string
        averageresult:
          type: number
        class:
          type: number
        date:
          type: integer
        lessonno:
          type: number
        maxmarks:
          type: number
        period:
          type: number
        subjectname:
          type: string
        teachername:
          type: string
      required:
        -  _id
        - class
        - date
        - lessonno
        - subjectname
        - teachername
    User:
      properties:
        _id:
          type: string
        mail:
          type: string
        name:
          type: string
        password:
          type: string
        roles:
          type: string
        surname:
          type: string
        username:
          type: string
      required:
        -  _id
        - password
        - username